pyproject.toml
[project]
name = "mcp-dataset-scalyr"
version = "0.1.0"
description = "mcp-dataset-scalyr MCP server built with FastMCP"
authors = [
    {name = "prasta", email = "prastamaha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
"mcp-dataset-scalyr-server" = "src.main:main"
dev = "src.main:dev"
dev-http = "src.main:dev"
start = "src.main:start"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 